Conflicts:
	client/lib/features/community/presentation/widgets/carousel/carousel_tabs.dart
	client/lib/features/events/features/event_page/presentation/event_tab_controller.dart
	client/lib/l10n/app_en.arb
	client/lib/l10n/app_es.arb
	client/lib/l10n/app_zh.arb
	client/lib/l10n/app_zh_Hant_TW.arb

+++ client/lib/features/announcements/presentation/views/announcments.dart
@@ -71,7 +71,7 @@ class _AnnouncementsState extends State<Announcements> {
                     cancelText: context.l10n.no,
                   ).show();
 
-                  if (confirmedDelete) {
+                  if (confirmedDelete && mounted) {
                     await firestoreAnnouncementsService.deleteAnnouncement(
                       communityId: CommunityProvider.read(context).communityId,
                       announcementId: announcement.id!,
