+++ client/lib/features/events/features/live_meeting/features/meeting_agenda/presentation/views/agenda_item_card.dart
@@ -363,7 +363,7 @@ class _AgendaItemCardState extends State<AgendaItemCard>
       cancelText: context.l10n.cancel,
     ).show(context: context);
 
-    if (delete) {
+    if (delete && mounted) {
       await alertOnError(context, () => _presenter.deleteAgendaItem());
     }
   }
@@ -391,7 +391,7 @@ class _AgendaItemCardState extends State<AgendaItemCard>
           onPressed: () async {
             if (hasBeenEdited) {
               final isDiscardChangesConfirmed = await ConfirmDialog(
-                mainText: 'Are you sure you want to discard changes?',
+                mainText: context.l10n.confirmDiscardChanges,
                 cancelText: context.l10n.cancel,
               ).show(context: context);
